feat: advanced todo example w/ pending and failed
dom/on= now returns a successed and failed atom. This is simpler than
returning deltas and force the user to re-create the atom. The overhead
of managing the atoms is visible in the advanced todo example where each
call needs to consume the atom's values in some way. It is also
error-prone for the user, one needs to remember to capture and consume
them otherwise exceptions might never print and memory might get
consumed.

Still, this is additional functionality we didn't have before.